treehacks-backend/requirements.txt
fastapi>=0.109.2
pydantic>=2.6.1
pydantic-core>=2.16.2
uvicorn>=0.27.1
requests==2.31.0
supabase>=2.13.0
python-dotenv>=1.0.0
starlette>=0.36.3
typing-extensions>=4.9.0
annotated-types>=0.6.0
anyio>=4.2.0
sniffio>=1.3.0
websockets>=12.0
gdown==5.1.0
python-multipart==0.0.6
httpx==0.27.0
opencv-python==4.9.0.80
pdf2image==1.17.0
numpy==1.26.3
google-generativeai
openai
